Sigh. Switch to slime for mgl-pax.
I'm well and truly annoyed both that I've spent so
much time looking into this and that Slynk and Swank
are distinct systems. Surely the differences are not
so vast that having two incompatible wire protocols
is really necessary?

Anyway, I'm all in on mgl-pax and am not wedded to
any specific details of slime or sly. Either will suffice.